added iterEnd & mwfork option. Fixed time and snapshout output from int->float

+++ runtime/src/libraryInterface.cpp
@@ -33,7 +33,7 @@ float octree::getTheta()
 
 void octree::setTEnd(float tEndx)
 {
-    tEnd = (int)tEndx;  
+    tEnd = tEndx;  
 }
 
 float octree::getTEnd()
